User authentication and different accounts

+++ server/routers/flights.py
@@ -1,6 +1,8 @@
 from server.database import database
-from server.models import AirportModel, FlightModel
-from fastapi import APIRouter, HTTPException
+from server.models import AirportModel, ClassType, CustomModel, FlightModel, SeatType, User
+from server.auth.users import get_current_user
+
+from fastapi import APIRouter, Depends, HTTPException
 from enum import Enum
 import datetime
 import math
@@ -21,6 +23,13 @@ class Sort(str, Enum):
     TICKET_CLASS = "ticket_class"
     DURATION = "duration"
     DISTANCE = "distance"
+
+async def check_flight_authorization(id: int, user: User) -> None:
+    res = database.execute_read_query(f"SELECT username FROM flights WHERE id = ?;", [id])
+    flight_username = res[0][0]
+
+    if flight_username != user.username:
+        raise HTTPException(status_code=403, detail="You are not authorized to modify this flight")
 
 # https://en.wikipedia.org/wiki/Haversine_formula
 def spherical_distance(origin: AirportModel, destination: AirportModel) -> int:
@@ -49,7 +58,7 @@ def spherical_distance(origin: AirportModel, destination: AirportModel) -> int:
     return round(distance);
 
 @router.post("", status_code=201)
-async def add_flight(flight: FlightModel) -> int:
+async def add_flight(flight: FlightModel, user: User = Depends(get_current_user)) -> int:
     if not (flight.date and flight.origin and flight.destination):
         raise HTTPException(status_code=404, 
                             detail="Insufficient flight data. Date, Origin, and Destination are required")
@@ -91,7 +100,7 @@ async def add_flight(flight: FlightModel) -> int:
 
         flight.duration = round(delta_minutes)
 
-    columns = FlightModel.get_attributes(False)
+    columns = FlightModel.get_attributes(ignore=["id"])
 
     query = "INSERT INTO flights ("
     for attr in columns:
@@ -101,18 +110,41 @@ async def add_flight(flight: FlightModel) -> int:
     query = query[:-1]
     query += ") RETURNING id;"
 
-    values = flight.get_values()
+    # only admins may add flights for other users
+    if flight.username and not user.is_admin:
+        raise HTTPException(status_code=403, detail="Only admins can add flights for other users")
+
+    explicit = {"username": user.username} if not flight.username else {}
+    values = flight.get_values(ignore=["id"], explicit=explicit)
 
     return database.execute_query(query, values)
 
+class FlightPatchModel(CustomModel):
+    date:           datetime.date|None = None
+    origin:         AirportModel|str|None = None
+    destination:    AirportModel|str|None = None
+    departure_time: str|None = None
+    arrival_time:   str|None = None
+    arrival_date:   datetime.date|None = None
+    seat:           SeatType|None = None
+    ticket_class:   ClassType|None = None
+    duration:       int|None = None
+    distance:       int|None = None
+    airplane:       str|None = None
+    flight_number:  str|None = None
+    notes:          str|None = None
 @router.patch("", status_code=200)
-async def update_flight(id: int, new_flight: FlightModel) -> int:
+async def update_flight(id: int, 
+                        new_flight: FlightPatchModel,
+                        user: User = Depends(get_current_user)) -> int:
+    await check_flight_authorization(id, user)
+
     if new_flight.empty():
         return id
 
     query = "UPDATE flights SET "
  
-    for attr in FlightModel.get_attributes(False):
+    for attr in FlightPatchModel.get_attributes():
         value = getattr(new_flight, attr)
         if value:
             query += f"{attr}=?," if value else ""
@@ -127,7 +159,9 @@ async def update_flight(id: int, new_flight: FlightModel) -> int:
     return database.execute_query(query, values)
 
 @router.delete("", status_code=200)
-async def delete_flight(id: int) -> int:
+async def delete_flight(id: int, user: User = Depends(get_current_user)) -> int:
+    await check_flight_authorization(id, user)
+
     return database.execute_query(
         """
         DELETE FROM flights WHERE id = ? RETURNING id;
@@ -143,13 +177,17 @@ async def get_flights(id: int|None = None,
                       order: Order = Order.DESCENDING,
                       sort: Sort = Sort.DATE,
                       start: datetime.date|None = None,
-                      end: datetime.date|None = None) -> list[FlightModel]|FlightModel:
+                      end: datetime.date|None = None,
+                      username: str|None = None,
+                      user: User = Depends(get_current_user)) -> list[FlightModel]|FlightModel:
 
-    id_filter = f"WHERE f.id = {str(id)}" if id else ""
+    user_filter = f"AND f.username = '{user.username}'" if not id else ""
+    if username:
+        user_filter = f"AND f.username = '{username}'"
 
-    date_filter_start = "WHERE" if not id and (start or end) else "AND" if start or end else ""
+    id_filter = f"AND f.id = {str(id)}" if id else ""
 
-    date_filter = ""
+    date_filter = "AND" if start or end else ""
     date_filter += f"JULIANDAY(date) > JULIANDAY('{start}')" if start else ""
     date_filter += " AND " if start and end else ""
     date_filter += f"JULIANDAY(date) < JULIANDAY('{end}')" if end else ""
@@ -159,11 +197,13 @@ async def get_flights(id: int|None = None,
             f.*,
             o.*, 
             d.*
-        FROM flights f 
-        JOIN airports o ON LOWER(f.origin) = LOWER(o.icao) 
-        JOIN airports d ON LOWER(f.destination) = LOWER(d.icao)
+        FROM flights f
+        JOIN airports o ON UPPER(f.origin) = o.icao
+        JOIN airports d ON UPPER(f.destination) = d.icao
+        WHERE 1=1
+        {user_filter}
         {id_filter}
-        {date_filter_start} {date_filter}
+        {date_filter}
         ORDER BY f.{sort.value} {order.value}
         LIMIT {limit}
         OFFSET {offset};"""
@@ -171,7 +211,9 @@ async def get_flights(id: int|None = None,
     res = database.execute_read_query(query);
 
     # get rid of origin, destination ICAOs for proper conversion
-    res = [ flight_db[:2] + flight_db[4:] for flight_db in res ]
+    # after this, each flight_db is in the format:
+    # [id, username, date, departure_time, ..., AirportModel, AirportModel]
+    res = [ flight_db[:3] + flight_db[5:] for flight_db in res ]
 
     flights = []
 
@@ -193,7 +235,7 @@ async def get_flights(id: int|None = None,
         flights.append(flight)
 
     if id and not flights:
-        raise HTTPException(status_code=404, detail=f"Flight with id '{str(id)}' not found.")
+        raise HTTPException(status_code=404, detail=f"Flight not found.")
 
     if id:
         return FlightModel.model_validate(flights[0])
